+++ src/main/kotlin/de/bdr/servko/keycloak/gematik/idp/rest/GematikIDPResource.kt
@@ -1,12 +1,13 @@
 /*
- *  Copyright 2023 Bundesdruckerei GmbH and/or its affiliates
- *  and other contributors.
+ * Copyright 2025 Bundesdruckerei GmbH and/or its affiliates
+ * and other contributors.
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
  *  You may obtain a copy of the License at
  *
  *  http://www.apache.org/licenses/LICENSE-2.0
+ *
  *
  *  Unless required by applicable law or agreed to in writing, software
  *  distributed under the License is distributed on an "AS IS" BASIS,
@@ -204,6 +205,7 @@ abstract class GematikIDPResource {
             GematikIDPUtil.getEndpointUri(session, realm, brokerState, config, GematikIdpLiterals.AUTHENTICATION_STATUS)
 
         loginFormsProvider.setAttribute("statusUrl", statusUrl)
+        loginFormsProvider.setAttribute(GematikIDPConfig.AUTHENTICATOR_AUTO_LAUNCH, config.getAuthenticatorAutoLaunch())
 
         return loginFormsProvider.createForm("gematik-idp.ftl")
     }